backend/requirements.txt
Django==5.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.5.1
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
